(BKR-273) Allow Setting Type by Host
- no longer use 'type' to control path/default host settings
- set environment and defaults post pe/foss installation
- dump more information post testing so that you can correctly pick up
  the host again
