+++ lib/beaker/answers.rb
@@ -3,6 +3,40 @@ module Beaker
   # information.
   class Answers
 
+    DEFAULT_ANSWERS =  Beaker::Options::OptionsHash.new.merge({
+      :q_puppet_enterpriseconsole_auth_user_email    => 'admin@example.com',
+      :q_puppet_enterpriseconsole_auth_password      => '~!@#$%^*-/ aZ',
+      :q_puppet_enterpriseconsole_smtp_port          => 25,
+      :q_puppet_enterpriseconsole_smtp_use_tls       => 'n',
+      :q_verify_packages                             => 'y',
+      :q_puppetdb_password                           => '~!@#$%^*-/ aZ',
+      :q_puppetmaster_enterpriseconsole_port         => 443,
+      :q_puppet_enterpriseconsole_auth_database_name => 'console_auth',
+      :q_puppet_enterpriseconsole_auth_database_user => 'mYu7hu3r',
+      :q_puppet_enterpriseconsole_database_name      => 'console',
+      :q_puppet_enterpriseconsole_database_user      => 'mYc0nS03u3r',
+      :q_database_root_password                      => '=ZYdjiP3jCwV5eo9s1MBd',
+      :q_database_root_user                          => 'pe-postgres',
+      :q_database_export_dir                         => '/tmp',
+      :q_puppetdb_database_name                      => 'pe-puppetdb',
+      :q_puppetdb_database_user                      => 'mYpdBu3r',
+      :q_database_port                               => 5432,
+      :q_puppetdb_port                               => 8081,
+      :q_classifier_database_user                    => 'DFGhjlkj',
+      :q_database_name                               => 'pe-classifier',
+      :q_classifier_database_password                => '~!@#$%^*-/ aZ',
+      :q_activity_database_user                      => 'adsfglkj',
+      :q_activity_database_name                      => 'pe-activity',
+      :q_activity_database_password                  => '~!@#$%^*-/ aZ',
+      :q_rbac_database_user                          => 'RbhNBklm',
+      :q_rbac_database_name                          => 'pe-rbac',
+      :q_rbac_database_password                      => '~!@#$%^*-/ aZ',
+      :q_install_update_server                       => 'y',
+      :q_exit_for_nc_migrate                         => 'n',
+      :q_enable_future_parser                        => 'n',
+      :q_pe_check_for_updates                        => 'n',
+    })
+
     # When given a Puppet Enterprise version, a list of hosts and other
     # qualifying data this method will return the appropriate object that can be used
     # to generate answer file data.
@@ -46,7 +80,16 @@ module Beaker
     # @param [String] default Should there be no user value for the provided question name return this default
     # @return [String] The answer value
     def answer_for(options, q, default = nil)
-      options[:answers][q] ? options[:answers][q] : default
+      answer = DEFAULT_ANSWERS[q]
+      # check to see if there is a value for this in the provided options
+      if options[:answers] && options[:answers][q]
+        answer = options[:answers][q]
+      end
+      # use the default if we don't have anything
+      if not answer
+        answer = default
+      end
+      answer
     end
 
     # When given a Puppet Enterprise version, a list of hosts and other
