+++ lib/beaker/answers/version30.rb
@@ -21,9 +21,9 @@ module Beaker
       agent_a = {
         :q_puppetagent_install => 'y',
         :q_puppet_cloud_install => 'y',
-        :q_verify_packages => options[:answers][:q_verify_packages],
+        :q_verify_packages => answer_for(options, :q_verify_packages, 'y'),
         :q_puppet_symlinks_install => 'y',
-        :q_puppetagent_certname => host,
+        :q_puppetagent_certname => host.to_s,
 
         # Disable database, console, and master by default
         # This will be overridden by other blocks being merged in.
@@ -33,7 +33,7 @@ module Beaker
         :q_puppetdb_install => 'n',
         :q_database_install => 'n',
       }
-      agent_a[:q_puppetagent_server] = masterless ? host : master
+      agent_a[:q_puppetagent_server] = masterless ? host.to_s : master.to_s
       agent_a[:q_continue_or_reenter_master_hostname] = 'c' if masterless
 
       # These base answers are needed by all
@@ -45,12 +45,12 @@ module Beaker
       unless masterless
         # master/database answers
         master_database_a = {
-          :q_puppetmaster_certname => master
+          :q_puppetmaster_certname => master.to_s
         }
 
         # Master/dashboard answers
         master_console_a = {
-          :q_puppetdb_hostname  => answer_for(options, :q_puppetdb_hostname, database),
+          :q_puppetdb_hostname  => answer_for(options, :q_puppetdb_hostname, database.to_s),
           :q_puppetdb_port      => answer_for(options, :q_puppetdb_port, 8081)
         }
 
@@ -59,15 +59,15 @@ module Beaker
         master_a = {
           :q_puppetmaster_install => 'y',
           :q_puppetmaster_dnsaltnames => master_dns_altnames,
-          :q_puppetmaster_enterpriseconsole_hostname => dashboard,
+          :q_puppetmaster_enterpriseconsole_hostname => dashboard.to_s,
           :q_puppetmaster_enterpriseconsole_port => answer_for(options, :q_puppetmaster_enterpriseconsole_port, 443),
         }
 
         # Common answers for console and database
         database_name = answer_for(options, :q_puppetdb_database_name, 'pe-puppetdb')
         database_user = answer_for(options, :q_puppetdb_database_user, 'mYpdBu3r')
-        dashboard_password = "'#{options[:answers][:q_puppet_enterpriseconsole_auth_password]}'"
-        puppetdb_password = "'#{options[:answers][:q_puppetdb_password]}'"
+        dashboard_password = "'#{answer_for(options, :q_puppet_enterpriseconsole_auth_password)}'"
+        puppetdb_password = "'#{answer_for(options, :q_puppetdb_password)}'"
         auth_database_name = answer_for(options, :q_puppet_enterpriseconsole_auth_database_name, 'console_auth')
         auth_database_user = answer_for(options, :q_puppet_enterpriseconsole_auth_database_user, 'mYu7hu3r')
         console_database_name = answer_for(options, :q_puppet_enterpriseconsole_database_name, 'console')
@@ -85,29 +85,29 @@ module Beaker
           :q_puppet_enterpriseconsole_database_user => answer_for(options, :q_puppet_enterpriseconsole_database_user, console_database_user),
           :q_puppet_enterpriseconsole_database_password => answer_for(options, :q_puppet_enterpriseconsole_database_password, dashboard_password),
 
-          :q_database_host => answer_for(options, :q_database_host, database),
+          :q_database_host => answer_for(options, :q_database_host, database.to_s),
           :q_database_port => database_port,
         }
 
         # Console only answers
-        dashboard_user = "'#{options[:answers][:q_puppet_enterpriseconsole_auth_user_email]}'"
+        dashboard_user = "'#{answer_for(options, :q_puppet_enterpriseconsole_auth_user_email)}'"
 
 
-        smtp_host = "'#{options[:answers][:q_puppet_enterpriseconsole_smtp_host] || dashboard}'"
-        smtp_port = "'#{options[:answers][:q_puppet_enterpriseconsole_smtp_port]}'"
-        smtp_username = options[:answers][:q_puppet_enterpriseconsole_smtp_username]
-        smtp_password = options[:answers][:q_puppet_enterpriseconsole_smtp_password]
-        smtp_use_tls = "'#{options[:answers][:q_puppet_enterpriseconsole_smtp_use_tls]}'"
+        smtp_host = "'#{answer_for(options, :q_puppet_enterpriseconsole_smtp_host, dashboard.to_s)}'"
+        smtp_port = "'#{answer_for(options, :q_puppet_enterpriseconsole_smtp_port)}'"
+        smtp_username = answer_for(options, :q_puppet_enterpriseconsole_smtp_username)
+        smtp_password = answer_for(options, :q_puppet_enterpriseconsole_smtp_password)
+        smtp_use_tls = "'#{answer_for(options, :q_puppet_enterpriseconsole_smtp_use_tls)}'"
         console_inventory_port = answer_for(options, :q_puppet_enterpriseconsole_inventory_port, 8140)
         console_httpd_port = answer_for(options, :q_puppet_enterpriseconsole_httpd_port, 443)
 
         console_a = {
           :q_puppet_enterpriseconsole_install => 'y',
-          :q_puppet_enterpriseconsole_inventory_hostname => host,
-          :q_puppet_enterpriseconsole_inventory_certname => host,
-          :q_puppet_enterpriseconsole_inventory_dnsaltnames => dashboard,
+          :q_puppet_enterpriseconsole_inventory_hostname => host.to_s,
+          :q_puppet_enterpriseconsole_inventory_certname => host.to_s,
+          :q_puppet_enterpriseconsole_inventory_dnsaltnames => dashboard.to_s,
           :q_puppet_enterpriseconsole_inventory_port => console_inventory_port,
-          :q_puppet_enterpriseconsole_master_hostname => master,
+          :q_puppet_enterpriseconsole_master_hostname => master.to_s,
 
           :q_puppet_enterpriseconsole_auth_user_email => dashboard_user,
           :q_puppet_enterpriseconsole_auth_password => dashboard_password,
