+++ lib/beaker/dsl/install_utils.rb
@@ -1,4 +1,4 @@
-[ 'pe', 'puppet', 'ezbake', 'module' ].each do |lib|
+[ 'foss', 'pe', 'puppet', 'ezbake', 'module' ].each do |lib|
     require "beaker/dsl/install_utils/#{lib}_utils"
 end
 
@@ -6,8 +6,9 @@ module Beaker
   module DSL
     # Collection of installation methods and support
     module InstallUtils
-      include DSL::InstallUtils::PEUtils
       include DSL::InstallUtils::PuppetUtils
+      include DSL::InstallUtils::PEUtils
+      include DSL::InstallUtils::FOSSUtils
       include DSL::InstallUtils::ModuleUtils
       include DSL::InstallUtils::EZBakeUtils
     end
