+++ lib/beaker/hypervisor/solaris.rb
@@ -24,7 +24,7 @@ module Beaker
       }
 
       @logger.notify "Connecting to hypervisor at #{hypername}"
-      hypervisor = Beaker::Host.create( hypername, hyperopts )
+      hypervisor = Beaker::Host.create( hypername, hyperopts, @options )
       hypervisor[:user] = @fog_file[:default][:solaris_hypervisor_username] || hypervisor[:user]
       hypervisor[:ssh][:keys] = [@fog_file[:default][:solaris_hypervisor_keyfile]] || hypervisor[:ssh][:keys]
