+++ lib/beaker/options/command_line_parser.rb
@@ -27,13 +27,6 @@ module Beaker
             @cmd_options[:options_file] =  file
           end
 
-          opts.on '--type TYPE',
-                  'one of git, foss, or pe',
-                  'used to determine underlying path structure of puppet install',
-                  '(default pe)' do |type|
-            @cmd_options[:type] = type
-          end
-
           opts.on '--helper PATH/TO/SCRIPT',
                   'Ruby file evaluated prior to tests',
                   '(a la spec_helper)' do |script|
@@ -193,6 +186,10 @@ module Beaker
             @cmd_options[:validate] = bool
           end
 
+          opts.on '--collect-perf-data', 'Use sysstat on linux hosts to collect performance and load data' do
+            @cmd_options[:collect_perf_data] = true
+          end
+
           opts.on('--version', 'Report currently running version of beaker' ) do
             @cmd_options[:version] = true
           end
@@ -220,9 +217,12 @@ module Beaker
             #noop
           end
 
-          opts.on '--collect-perf-data', 'Use sysstat on linux hosts to collect performance and load data' do
-            @cmd_options[:collect_perf_data] = true
+          opts.on '--type TYPE',
+                  'DEPRECATED - pe/foss/aio determined during runtime' do |type|
+            #backwards compatability, oh how i hate you
+            @cmd_options[:type] = type
           end
+
         end
 
       end
