+++ lib/beaker/options/presets.rb
@@ -165,39 +165,6 @@ module Beaker
           :host_env               => {},
           :ssh_env_file           => '~/.ssh/environment',
           :profile_d_env_file     => '/etc/profile.d/beaker_env.sh',
-          :answers                => {
-                                     :q_puppet_enterpriseconsole_auth_user_email    => 'admin@example.com',
-                                     :q_puppet_enterpriseconsole_auth_password      => '~!@#$%^*-/ aZ',
-                                     :q_puppet_enterpriseconsole_smtp_port          => 25,
-                                     :q_puppet_enterpriseconsole_smtp_use_tls       => 'n',
-                                     :q_verify_packages                             => 'y',
-                                     :q_puppetdb_password                           => '~!@#$%^*-/ aZ',
-                                     :q_puppetmaster_enterpriseconsole_port         => 443,
-                                     :q_puppet_enterpriseconsole_auth_database_name => 'console_auth',
-                                     :q_puppet_enterpriseconsole_auth_database_user => 'mYu7hu3r',
-                                     :q_puppet_enterpriseconsole_database_name      => 'console',
-                                     :q_puppet_enterpriseconsole_database_user      => 'mYc0nS03u3r',
-                                     :q_database_root_password                      => '=ZYdjiP3jCwV5eo9s1MBd',
-                                     :q_database_root_user                          => 'pe-postgres',
-                                     :q_database_export_dir                         => '/tmp',
-                                     :q_puppetdb_database_name                      => 'pe-puppetdb',
-                                     :q_puppetdb_database_user                      => 'mYpdBu3r',
-                                     :q_database_port                               => 5432,
-                                     :q_puppetdb_port                               => 8081,
-                                     :q_classifier_database_user                    => 'DFGhjlkj',
-                                     :q_database_name                               => 'pe-classifier',
-                                     :q_classifier_database_password                => '~!@#$%^*-/ aZ',
-                                     :q_activity_database_user                      => 'adsfglkj',
-                                     :q_activity_database_name                      => 'pe-activity',
-                                     :q_activity_database_password                  => '~!@#$%^*-/ aZ',
-                                     :q_rbac_database_user                          => 'RbhNBklm',
-                                     :q_rbac_database_name                          => 'pe-rbac',
-                                     :q_rbac_database_password                      => '~!@#$%^*-/ aZ',
-                                     :q_install_update_server                       => 'y',
-                                     :q_exit_for_nc_migrate                         => 'n',
-                                     :q_enable_future_parser                        => 'n',
-                                     :q_pe_check_for_updates                        => 'n',
-          },
           :dot_fog                => File.join(ENV['HOME'], '.fog'),
           :ec2_yaml               => 'config/image_templates/ec2.yaml',
           :help                   => false,
