+++ spec/beaker/options/presets_spec.rb
@@ -19,32 +19,6 @@ module Beaker
         expect(presets.presets).to be_instance_of(Beaker::Options::OptionsHash)
       end
 
-      describe 'when setting the type as pe from the environment' do
-        describe 'sets type to pe if...' do
-          it 'env var is set to "true"' do
-            munged = presets.format_found_env_vars( {:is_pe => 'true'} )
-            expect( munged[:type] ).to be == 'pe'
-          end
-          it 'env var is set to "yes"' do
-            munged = presets.format_found_env_vars( {:is_pe => 'yes'} )
-            expect( munged[:type] ).to be == 'pe'
-          end
-        end
-        describe 'sets type to foss if...' do
-          it 'env var is set to "false"' do
-            munged = presets.format_found_env_vars( {:is_pe => 'true'} )
-            expect( munged[:type] ).to be == 'pe'
-          end
-          it 'env var is set to "no"' do
-            munged = presets.format_found_env_vars( {:is_pe => 'yes'} )
-            expect( munged[:type] ).to be == 'pe'
-          end
-        end
-        it 'does not set type otherwise' do
-          munged = presets.format_found_env_vars( {:is_pe => nil} )
-          expect( munged[:type] ).to be == nil
-        end
-      end
     end
   end
 end
